倒计时2.0  可能存在bug的情况修复

+++ app/src/main/java/net/tatans/rhea/countdowntimer/adapter/DelSchemeAdapter.java
@@ -114,11 +114,12 @@ public class DelSchemeAdapter extends BaseAdapter {
 
         @Override
         public void onClick(View v) {
+            if (al_countDown.size() == 0) return;
             bean = al_countDown.get(mPosition);
             tdb.delete(bean);
             al_countDown = tdb.findAll(CountDownBean.class);
             interrupt();
-            TatansToast.showShort("删除成功");
+            TatansToast.showAndCancel("删除成功");
             notifyDataSetChanged();
         }
     }
